requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
mysql-connector-python==8.0.33
vercel-wsgi==0.2.1
